move tool utils to common (#4081)
* move tool utils to common


* avoid adding spi dependency to common


* add javadoc to ToolUtils


---------
