+++ ml-algorithms/src/main/java/org/opensearch/ml/engine/tools/McpSseTool.java
@@ -15,6 +15,7 @@ import org.opensearch.ml.common.spi.tools.Parser;
 import org.opensearch.ml.common.spi.tools.ToolAnnotation;
 import org.opensearch.ml.common.spi.tools.WithModelTool;
 import org.opensearch.ml.common.utils.StringUtils;
+import org.opensearch.ml.common.utils.ToolUtils;
 import org.opensearch.ml.repackage.com.google.common.annotations.VisibleForTesting;
 
 import io.modelcontextprotocol.client.McpSyncClient;
